Ignore Shim6 control packets and ICMP in shim6_output
The previous technique was to create a special policy that prevented
shim6 control and ICMP packets from even going through shim6_output.
Due to the collab with MIPv6, it is not possible anymore to do so.
We thus removed the special policy in the daemon, and we put that
filter inside shim6_output.