tcp_probev6:Removed KPROBES dependency for uml compatibility+IPV6 check
Previously tcp_probev6 used KPROBES to insert a hook into the
tcp_rcv_established function. In order to be able to debug under UML,
We replaced this with an explicit call inside the tcp_rcv_established
function. It is a temporary solution, however, since it introduces
additional code in the main TCP code, which was avoided with KPROBES.

BUG FIX: We also check now if the received packet is IPv6. If not, it is
simply ignored.