Added ART switch handle in /proc/net/ipv6/shim6
This ART switch is almost always 0. But user space can set it to 1 at
any time, if it wants to inform the kernel that an event happened and
the ART should be measured. For example, if an application does a
movement, the correct behaviour is to trigger the movement, and
immediately after set the artswitch to 1. The kernel will set it again
to 0 as soon as the first packet is received, after the event, and will
send the measured time to the shim6d daemon. In turn, the shim6d daemon
will write the result in /etc/shim6/art.log.